Add zplug support and default installation

+++ zsh/install.macos.sh
@@ -1,9 +1,21 @@
+. "$DOTFILES/scripts/core/main.sh"
+
 # Check version exisiting
+if ! platform::command_exists "zsh"
+then
+  # Install zsh and zsh-completions
+  brew install zsh zsh-completions
+fi
 
-# Install zsh and zsh-completions
-brew install zsh zsh-completions
+# Set as default shell
+ZSH_SHELL_LOC=/bin/zsh
+if [ "$SHELL" != "$ZSH_SHELL_LOC" ]
+then
+  chsh -s /bin/zsh
+fi
 
-# Set as default
-chsh -s /bin/zsh
-
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+# Install zplug for the next bit
+if [[ -z $ZPLUG_HOME ]]; then
+  export ZPLUG_HOME=~/.zplug
+  git clone https://github.com/zplug/zplug $ZPLUG_HOME
+fi
